+++ providers/jdbc_resource.rb
@@ -35,6 +35,24 @@ action :create do
     group node['glassfish']['group']
     code asadmin_command(command.join(' '))
   end
+
+  sets = {'pool-name' => new_resource.connectionpoolid, 'description' => new_resource.description}
+  new_resource.properties.each_pair do |key, value|
+    sets["property.#{key}"] = value
+  end
+  sets['enabled'] = !!new_resource.enabled
+  sets.each_pair do |key, value|
+    variable = "resources.jdbc-resource.#{new_resource.name}.#{key}"
+    glassfish_property "#{variable}=#{value}" do
+      domain_name new_resource.domain_name
+      admin_port new_resource.admin_port
+      username new_resource.username
+      password_file new_resource.password_file
+      secure new_resource.secure
+      key variable
+      value value.to_s
+    end
+  end
 end
 
 action :delete do